Fix a crash caused by no-arg lambda or let_values.
There's an optimization to avoid creating an empty
layer of storage when there's no variables in a
rib. But that wasn't mirrored in the environment
creation during expand and compile, so variable
lookups that crossed the missing rib were going
too far up the lexical stack. This would either
return the wrong value or crash when the lookup
hit top-level.

FUSION-273