Fix defective symbol interning code.
Due to incorrect use of WeakHashMap, there could
be multiple simultaneous instances of ActualSymbol
with the same text content. The core error is that
WeakHashMap.put() does not guarantee that the key
argument will be retained by the map entry; it
could reuse the key that's already there. So when
the ActualSymbol has been released but the String
key is not, we end up violating the invariant that
the key == the ActualSymbol's text.

The fix is to stop trying to avoid the allocation
of an ActualSymbol, so that the map's keys and
values are the same instance and their respective
WeakReferences are cleared simultaneously.

Issue ID FUSION-47

Code review ID 5011927