Render docs for binding usage as part of the body div.
Since BindingDoc.usage is only populated via the deprecated "doc on value" code paths, different markup was being generated.  Additionally, the markup was wrong, since the 'oneliner' class was intended for a short textual description of the feature, not the usage example.

Now the usage example is part of the body span, regardless of how it was derived from the source code.