Port Racket's implementation of `for_fold`.
This provides a more extensible foundation for the comprehension forms.

I wanted to start adding more of Racket's [`for`-family of iteration forms](https://docs.racket-lang.org/reference/for.html) but quickly realized that the approach I'd taken for `for_list` and `fors_list` was not going to scale. Doing some research, I realized that it would be better to take Racket's approach by building up from `for/fold/derived` so all the iteration and collation logic can be reused. See the implementation notes in `for.fusion` for more details.

Code review ID CR-10491491