Fix non-failure of undefined forward reference at module-level.
https://docs.racket-lang.org/reference/module.html says:

> [E]xpressions and definitions are evaluated in order as they appear within the module.
>
> Accessing a module-level variable before it is defined signals a run-time error, just like accessing an undefined global variable.

Such references would to date cause a a JRE `AssertionError` at runtime, but only if assertions were enabled. Otherwise, the reference would result in void, somewhat by accident, since the `Evaluator` still (unfortunately) transforms all Java-null results to void.

Most of the complexity in this change is to persist the names of module-level definitions inside `NamespaceStore`, so that when an undefined reference is evaluated we can determine the name to display.

Thanks to Rob Calhoun for reporting the bug.

Code review ID CR-7354798