Fix a crash when assertions are disabled.
The removed assertion was inappropriate since it could have a side effect.

This code has been broken for a long time.  When assertions were disabled, the removed call to `getBinding()` was returning null (which is not expected) but that was compensated by the corresponding compiler code calling `resolve()`.  In general, expansion code should use `resolve()` to force the identifier to have a binding, and compiler code, as noted in the removed TODO, is expected to fast-track via `getBinding()`.

In other words, there were two latent bugs that "fixed each other", that no longer do so with the recently-improved caching in SyntaxSymbol, and this fact was hidden (in unit tests) by an assertion with a side-effect.

I've audited the code and no other assert calls `resolve()`.

Code review ID CR-19065251