Use the new MultiHashTrie from FusionStruct.
We have finally reach the change that I've been working toward since 2020-08: pulling the implementation of multi-hash out of the implementation of structs. This eliminates a ton of complexity from the latter, and makes the capability available to the rest of the implementation (including, perhaps, a user-facing multi-hash data structure).

The code could still use a cleanup pass to update some member and method names that are no longer accurate, but that's a different refactoring step for a different day.

Code review ID CR-61988327