Fix bug in lexical context of `require`.
Previously, forms that import a full module eg `(require "/module")`
take the lexical context from the `require` keyword and apply it to
the imported bindings. This is contrary to Racket's behavior, and makes
it impossible to write macros that introduce imports that affect the
code outside the macro invocation _and_ that work properly when
`require` is bound differently in that context.

Now we follow Racket's specification and use the lexical context from
the module-path sub-form.

Other require-spec forms have similar issues; they'll be addressed in
future commits.

Code review ID CR-80584418