Teach code coverage to handle URL SourceNames.
* Pass the URL through `CoverageDatabase` serialization.
* Handle such URLs in the coverage writer, extracting source code from
  the jar when rendering them to highlighted HTML.

This is not ideal.  Fully-general URL sources don't fit well with the assumptions of code coverage.  This may work better if it's tuned to Jar resources.

We should also replace or supplement the "path prefix" mechanism (used to shorted paths in the report) with explicit handling of repository roots.
Then we could treat each Jar as a root with independent reporting, which  would make more sense to the customer.

Code review ID CR-153632372