Remove special treatment of toplevel module IDs. (#181)
This removes some complex code that was needed to handle the divergent meaning of relative module paths between modules and top-levels (see #166). Now everything works the same, except for a small dispatch at the Namespace level to handle that difference.

We still need a way to create unresolvable paths that provide the "scoping" of module names that occurs at toplevel, and now that behavior is better abstracted.

As a side effect, the identities of toplevel namespaces are now interned as are all the rest.  Because interned IDs are never removed from the table, this exacerbates an existing memory leak in applications that churn toplevels.  This increases the priority of #164 (while fixing part of it).